fix: HASSU-1902 Korjaa kiinteistönomistajalistaukseen jääneitä ongelmia (#1089)
* fix: HASSU-1902 Korjaa kiinteistönomistajalistaukseen jääneitä ongelmia

* Update bankHolidaysClient.ts

* Lisää VIRHE_AIKAKATKAISTU OmistajahakuTila enumiin ja muokkaa logiikkaa tukemaan sitä

* poista epärelevantti kommentti

+++ backend/src/projektiSearch/omistajaSearch/omistajaDynamoDBStreamHandler.ts
@@ -38,9 +38,9 @@ async function handleRemove(record: DynamoDBRecord) {
 }
 
 async function handleManagementEvent(event: MaintenanceEvent) {
-  if (event.action == "deleteIndex") {
+  if (event.action === "deleteIndex") {
     await new OmistajaSearchMaintenanceService().deleteIndex();
-  } else if (event.action == "index") {
+  } else if (event.action === "index") {
     let startKey: OmistajaKey | undefined = undefined;
     const queueUrl = await parameters.getOmistajaIndexerSQSUrl();
     do {
@@ -66,7 +66,7 @@ export const handleDynamoDBEvents = async (event: Event): Promise<void> => {
   } else if (eventIsSqsEvent(event)) {
     await handleSqsEvent(event);
   } else {
-    handleStreamEvent(event);
+    await handleStreamEvent(event);
   }
 };
 
@@ -80,16 +80,17 @@ function handleStreamEvent(event: DynamoDBStreamEvent) {
     return (async () => {
       setupLambdaMonitoringMetaData(subsegment);
       try {
-        for (const record of event.Records) {
-          switch (record.eventName) {
-            case "INSERT":
-            case "MODIFY":
-              await handleUpdate(record);
-              break;
-            case "REMOVE":
-              await handleRemove(record);
+        const promises = event.Records.map(async (record) => {
+          if (!record.eventName) {
+            return;
           }
-        }
+          if (record.eventName === "INSERT" || record.eventName === "MODIFY") {
+            await handleUpdate(record);
+          } else if (record.eventName === "REMOVE") {
+            await handleRemove(record);
+          }
+        });
+        await Promise.all(promises);
       } catch (e: unknown) {
         log.error(e);
         throw e;
